x86 Long Mode Enabled
According to Qemu I have successfully put the processor into Long Mode.
Right now it has a simple 1GB identity map and a new GDT. It's currently
not a higher half kernel yet, that's the next step. Also many checks
need to be done using CPUID to figure out the maximum number of address
lines and whether or not Long Mode is supported at all.