Virtual Memory Sub-System Update
The physical allocator is coming along. It currently figures out all of
the valid physical RAM and then creates an identity mapping of it inside
of kernel space. The next step is to create the physical allocator.

The virtual allocator is not ready yet.

Stack is unfinished, but will eventually be the basis for the physical
memory allocator.