Fleshed out some more of the virtual memory manager.
Filled in the body of some of the functions. Hopefully fixed some of the
macros (untested). Some of the functionality will change and there are
most likely many bugs in the code because the copy-on-write semantics
haven't been thought of too much yet, only some initial planning. That
will be tackled once processes come in to play and it's actually
necessary.